[BUGFIX] Fix broken URL to "/" because of pages.l18n_cfg=1 or 2
Problem description: The sitemap renders also url to pages (because they existing) even if the current language does not allow the page to be rendered because of "Hide default translation of page" (pages.l18n_cfg).
Solution: Don't render those url.
I'm not totally sure what the "clean up" stands for but I think it can be removed. Even links to home are not "/" but with an absolute url like "https://domain.org/"